chore: hooks file renaming
