+++ ffxi-crafting-app/src/hooks/useApi.ts
